backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
google-auth
google-auth-oauthlib
google-api-python-client
sentence-transformers
chromadb
transformers
torch                # for CPU; if you have a GPU install the suitable torch wheel
python-jose[cryptography]
cryptography
pdfminer.six
python-docx
pypandoc
python-magic
aiofiles
requests
